+++ Sources/_Datadog_Private/include/ObjcAppLaunchHandler.h
@@ -4,7 +4,7 @@
 * Copyright 2019-2020 Datadog, Inc.
 */
 
-#import <CoreFoundation/CFDate.h>
+#import <CoreFoundation/CoreFoundation.h>
 
 /// Returns the time interval between startup of the application process and the
 /// `UIApplicationDidBecomeActiveNotification`.
@@ -12,3 +12,8 @@
 /// If the `UIApplicationDidBecomeActiveNotification` has not been reached yet,
 /// it returns  time interval between startup of the application process and now.
 CFTimeInterval __dd_private_AppLaunchTime(void);
+
+/// Returns `true` when the application is pre-warmed.
+///
+/// System sets environment variable `ActivePrewarm` to 1 when app is pre-warmed.
+BOOL __dd_private_isActivePrewarm(void);